Add files via upload
Class Assignment [#01]

The "index.html" file contains the the HTML forms based on the given assignment task. The file is linked by "styles.css", "script.js", and "submit.php"
The "styles.css" is a file to style the HTML design
The "script.js" file contains the regexes of each HTML form. It also contains the code where if the user inputs something that does not meet the requirement of the defined regexes, it will display an error message in the web page.
The "submit.php" is mainly to retrieve the input texts done by the user in the HTML form, if all the input matches the defined regexes in the "script.js" file